(DEPRECATED) Import Raman spectra/maps from Andor cameras/solis ASCII files

These data input functions are deprecated and they will be removed in the next release of hyperspec package. Now functions in package hySpc.read.txt (link) should be used as the alternatives.

read.asc.Andor(
  file = stop("filename or connection needed"),
  ...,
  quiet = TRUE,
  dec = ".",
  sep = ","
)

Arguments

file
filename or connection to ASCII file

..., quiet, dec, sep
handed to base::scan()

Value

a hyperSpec object

Details

read.asc.Andor() reads Andor Solis ASCII (.asc) files where the first column gives the wavelength axes and the other columns the spectra.

See also

Author

Claudia Beleites