(DEPRECATED) Read Shimadzu GCxGC-qMS - spectra files (.txt)

These data input functions are deprecated and they will be removed in the next release of hyperspec package. Now functions in package hySpc.read.txt (link) should be used as the alternatives.

read.txt.Shimadzu(filename, encoding = "", quiet = TRUE)

Arguments

filename
file name and path of the .txt file

encoding
encoding of the txt file (used by base::readLines())

quiet
suppress printing of progress

Value

list of spectra tables

Details

Reads Shimadzu GCxGC-qMS - Spectra Files (.txt) as exported by Shimadzu Chrome Solution (v. 2.72). Mass Spectrometer: Shimadzu GCMS-QP 2010 Ultra (www.shimadzu.com)

Note

This is a first rough import function and the functions may change without notice.

Author

Bjoern Egert