(DEPRECATED) Import Raman spectra from Witec files

These data input functions are deprecated and they will be removed in the next release of hyperspec package. Now functions in package hySpc.read.txt (link) should be used as the alternatives.

read.txt.Witec(
  file = NULL,
  points.per.line = NULL,
  lines.per.image = NULL,
  type = c("single", "map"),
  hdr.label = FALSE,
  hdr.units = FALSE,
  encoding = "unknown",
  ...,
  quiet = TRUE
)

read.dat.Witec(
  filex = NULL,
  filey = sub("-x", "-y", filex),
  points.per.line = NULL,
  lines.per.image = NULL,
  type = c("single", "map"),
  encoding = "unknown",
  ...,
  quiet = hy_get_option("debuglevel") < 1L
)

read.txt.Witec.Graph(
  headerfile = NULL,
  filex = gsub("Header", "X-Axis", headerfile),
  filey = gsub("Header", "Y-Axis", headerfile),
  type = c("single", "map"),
  encoding = "unknown",
  ...,
  quiet = TRUE
)

Arguments

file

filename or connection to ASCII file

points.per.line

number of spectra in x direction of the map

lines.per.image

number of spectra in y direction

type

type of spectra: single for single spectra (including time series), map for imaging data.

hdr.label

WITec Project exports the spectra names (contain information of map position or number of spectra) within the file.

hdr.units

WITec Project exports the spectra units within the file.

encoding

character encoding, see base::readLines()

..., quiet

handed to base::scan()

filex

filename wavelength axis file

filey

filename intensity file

headerfile

filename or connection to ASCII file with header information

Value

a hyperSpec object

Details

Import Raman Spectra/Maps from Witec Instrument via ASCII files

read.txt.Witec() reads Witec ASCII files where the first column gives the wavelength axes and the other columns the spectra.

read.dat.Witec() reads Witec's ASCII exported data, which comes in separate files with x and y data.

See also

Author

Claudia Beleites and Marcel Dahms