(DEPRECATED) Convert different wavelength units

These hyperSpec functions are deprecated and not maintained any more. You should not use these. Currently they are present due to back-compatibility reasons and will be removed in the next release of the package. Please, use the suggested alternative functions instead.

_____________

The following units can be converted into each other: nm, \(cm^{-1}\), eV, THz and Raman shift

wlconv(points, src, dst, laser = NULL)

nm2raman(x, laser)

nm2invcm(x, ...)

nm2ev(x, ...)

nm2freq(x, ...)

invcm2raman(x, laser)

invcm2nm(x, ...)

invcm2ev(x, ...)

invcm2freq(x, ...)

raman2invcm(x, laser)

raman2nm(x, laser)

raman2ev(x, laser)

raman2freq(x, laser)

ev2raman(x, laser)

ev2invcm(x, ...)

ev2nm(x, ...)

ev2freq(x, ...)

freq2nm(x, ...)

freq2invcm(x, ...)

freq2ev(x, ...)

freq2raman(x, laser)

Arguments

points
data for conversion

src
source unit

dst
destination unit

laser
laser wavelength (required for work with Raman shift)

x
wavelength points for conversion

...
ignored

Functions

nm2raman: convevrsion nanometers -> Raman shift (relative wavenumber)

nm2invcm: conversion nanometers -> inverse cm (absolute wavenumber)

nm2ev: conversion nanometers -> electronvolt

nm2freq: conversion nm -> frequency in THz

invcm2raman: conversion inverse cm (absolute wavenumber) -> Raman shift (relative wavenumber)

invcm2nm: conversion inverse cm (absolute wavenumber) -> nanometers

invcm2ev: conversion inverse cm (absolute wavenumber) -> electronvolt

invcm2freq: conversion inverse cm (absolute wavenumber) -> frequency in THz

raman2invcm: conversion Raman shift (relative wavenumber) -> inverse cm (absolute wavenumber)

raman2nm: conversion Raman shift (relative wavenumber) -> nanometers

raman2ev: conversion Raman shift (relative wavenumber) -> electronvolt

raman2freq: conversion Raman shift (relative wavenumber) -> frequency in THz

ev2raman: conversion electronvolt -> Raman shift (relative wavenumber)

ev2invcm: conversion electronvolt -> inverse cm (absolute wavenumber)

ev2nm: conversion electronvolt -> nanometers

ev2freq: conversion electronvolt -> frequency in THz

freq2nm: conversion frequency in THz -> nanometers

freq2invcm: conversion frequency in THz -> inverse cm (absolute wavenumber)

freq2ev: conversion frequency in THz -> electronvolt

freq2raman: conversion frequency in THz -> Raman shift (relative wavenumber)

Author

R. Kiselev

Examples

wlconv(3200, "Raman shift", "nm", laser = 785.04)

#> Warning: Function 'wlconv' is deprecated. 
#> Use function 'wl_convert_units' instead.

#> Warning: Function 'f' is deprecated. 
#> Use function 'wl_convert_units' instead.

#> [1] 1048.415

wlconv(785, "nm", "invcm")

#> Warning: Function 'wlconv' is deprecated. 
#> Use function 'wl_convert_units' instead.

#> Warning: Function 'f' is deprecated. 
#> Use function 'wl_convert_units' instead.

#> [1] 12738.85