Binding hyperSpec Objects

The former difficulties with binding S4 objects are resolved since R version 3.2.0 and cbind and rbind now work as intended and expected for hyperSpec objects.

cbind2 binds the spectral matrices of two hyperSpec objects by column. All columns besides spc with the same name in x@data and y@data must have the same elements. Rows are ordered before checking.

bind(
  direction = stop("direction ('c' or 'r') required"),
  ...,
  wl.tolerance = hy.getOption("wl.tolerance")
)

# S3 method for hyperSpec
cbind(...)

# S3 method for hyperSpec
rbind(...)

# S4 method for hyperSpec,hyperSpec
cbind2(x, y)

# S4 method for hyperSpec,missing
cbind2(x, y)

# S4 method for hyperSpec,hyperSpec
rbind2(x, y, wl.tolerance = hy.getOption("wl.tolerance"))

# S4 method for hyperSpec,missing
rbind2(x, y, wl.tolerance)

Arguments

direction
"r" or "c" to bind rows or columns

...
The hyperSpec objects to be combined.

Alternatively, one list of hyperSpec objects can be given to bind.

wl.tolerance
rbind and rbind2 check for equal wavelengths with this tolerance.

x, y
hyperSpec objects

Value

a hyperSpec object, possibly with different row order (for bind ("c", ...{}) and cbind2).

Details

Therefore, calling rbind.hyperSpec and cbind.hyperSpec is now depecated: cbind and rbind should now be called directly.

However, in consequence it is no longer possible to call cbind or rbind with a list of hyperSpec objects. In that case, use bind or do.call (see example).

bind does the common work for both column- and row-wise binding.

Note

You might have to make sure that the objects either all have or all do not have rownames and/or colnames.

See also

Author

C. Beleites

Examples

chondro

#> hyperSpec object
#>    875 spectra
#>    5 data columns
#>    300 data points / spectrum
#> wavelength: Delta * tilde(nu)/cm^-1 [numeric] 602 606 ... 1798 
#> data:  (875 rows x 5 columns)
#>    1. y: y [numeric] -4.77 -4.77 ... 19.23 
#>    2. x: x [numeric] -11.55 -10.55 ... 22.45 
#>    3. filename: filename [character] rawdata/chondro.txt rawdata/chondro.txt ... rawdata/chondro.txt 
#>    4. clusters: clusters [factor] matrix matrix ... lacuna + NA
#>    5. spc: I / a.u. [matrix, array300] 501.8194 500.4552 ... 169.2942 

bind ("r", chondro, chondro)

#> hyperSpec object
#>    1750 spectra
#>    5 data columns
#>    300 data points / spectrum
#> wavelength: Delta * tilde(nu)/cm^-1 [numeric] 602 606 ... 1798 
#> data:  (1750 rows x 5 columns)
#>    1. y: y [numeric] -4.77 -4.77 ... 19.23 
#>    2. x: x [numeric] -11.55 -10.55 ... 22.45 
#>    3. filename: filename [character] rawdata/chondro.txt rawdata/chondro.txt ... rawdata/chondro.txt 
#>    4. clusters: clusters [factor] matrix matrix ... lacuna + NA
#>    5. spc: I / a.u. [matrix, array300] 501.8194 500.4552 ... 169.2942 

rbind (chondro, chondro)

#> hyperSpec object
#>    1750 spectra
#>    5 data columns
#>    300 data points / spectrum
#> wavelength: Delta * tilde(nu)/cm^-1 [numeric] 602 606 ... 1798 
#> data:  (1750 rows x 5 columns)
#>    1. y: y [numeric] -4.77 -4.77 ... 19.23 
#>    2. x: x [numeric] -11.55 -10.55 ... 22.45 
#>    3. filename: filename [character] rawdata/chondro.txt rawdata/chondro.txt ... rawdata/chondro.txt 
#>    4. clusters: clusters [factor] matrix matrix ... lacuna + NA
#>    5. spc: I / a.u. [matrix, array300] 501.8194 500.4552 ... 169.2942 

cbind (chondro, chondro)

#> hyperSpec object
#>    875 spectra
#>    5 data columns
#>    600 data points / spectrum
#> wavelength: Delta * tilde(nu)/cm^-1 [numeric] 602 606 ... 1798 
#> data:  (875 rows x 5 columns)
#>    1. y: y [numeric] -4.77 -4.77 ... 19.23 
#>    2. x: x [numeric] -11.55 -10.55 ... 22.45 
#>    3. filename: filename [character] rawdata/chondro.txt rawdata/chondro.txt ... rawdata/chondro.txt 
#>    4. clusters: clusters [factor] matrix matrix ... lacuna + NA
#>    5. spc: I / a.u. [matrix, array600] 501.8194 500.4552 ... 169.2942 

bind ("r", list (chondro, chondro, chondro))

#> hyperSpec object
#>    2625 spectra
#>    5 data columns
#>    300 data points / spectrum
#> wavelength: Delta * tilde(nu)/cm^-1 [numeric] 602 606 ... 1798 
#> data:  (2625 rows x 5 columns)
#>    1. y: y [numeric] -4.77 -4.77 ... 19.23 
#>    2. x: x [numeric] -11.55 -10.55 ... 22.45 
#>    3. filename: filename [character] rawdata/chondro.txt rawdata/chondro.txt ... rawdata/chondro.txt 
#>    4. clusters: clusters [factor] matrix matrix ... lacuna + NA
#>    5. spc: I / a.u. [matrix, array300] 501.8194 500.4552 ... 169.2942 

x <- chondro[,, 600 : 605]
x$a <- 1
x@data <- x@data[, sample (ncol (x), ncol (x))] # reorder columns

y <- chondro [nrow (chondro) : 1,, 1730 : 1750] # reorder rows
y$b <- 2

cbind2 (x, y) # works

#> hyperSpec object
#>    875 spectra
#>    7 data columns
#>    27 data points / spectrum
#> wavelength: Delta * tilde(nu)/cm^-1 [numeric] NA NA ... 1750 + NA
#> data:  (875 rows x 7 columns)
#>    1. a: a [numeric] 1 1 ... 1 
#>    2. y: y [numeric] -4.77 -4.77 ... 19.23 
#>    3. spc: I / a.u. [matrix, array27] NA NA ... 295.8961 + NA
#>    4. filename: filename [character] rawdata/chondro.txt rawdata/chondro.txt ... rawdata/chondro.txt 
#>    5. clusters: clusters [factor] matrix matrix ... lacuna + NA
#>    6. x: x [numeric] -11.55 -10.55 ... 22.45 
#>    7. b:  [numeric] 2 2 ... 2 

y$y[3] <- 5
try (cbind2 (x, y)) # error

#> hyperSpec object
#>    875 spectra
#>    7 data columns
#>    27 data points / spectrum
#> wavelength: Delta * tilde(nu)/cm^-1 [numeric] NA NA ... 1750 + NA
#> data:  (875 rows x 7 columns)
#>    1. a: a [numeric] 1 1 ... 1 
#>    2. y: y [numeric] -4.77 -4.77 ... 19.23 
#>    3. spc: I / a.u. [matrix, array27] NA NA ... 295.8961 + NA
#>    4. filename: filename [character] rawdata/chondro.txt rawdata/chondro.txt ... rawdata/chondro.txt 
#>    5. clusters: clusters [factor] matrix matrix ... lacuna + NA
#>    6. x: x [numeric] -11.55 -10.55 ... 22.45 
#>    7. b:  [numeric] 2 2 ... 2 

# list of hyperSpec objects

lhy <- list (flu, flu)
do.call ("rbind", lhy)

#> hyperSpec object
#>    12 spectra
#>    3 data columns
#>    181 data points / spectrum
#> wavelength: lambda/nm [numeric] 405.0 405.5 ... 495 
#> data:  (12 rows x 3 columns)
#>    1. spc: I[fl]/"a.u." [matrix, array181] 27.15000 66.80133 ... 294.6495 
#>    2. filename: filename [character] rawdata/flu1.txt rawdata/flu2.txt ... rawdata/flu6.txt 
#>    3. c: c / (mg / l) [numeric] 0.05 0.10 ... 0.3 

bind ("r", lhy)

#> hyperSpec object
#>    12 spectra
#>    3 data columns
#>    181 data points / spectrum
#> wavelength: lambda/nm [numeric] 405.0 405.5 ... 495 
#> data:  (12 rows x 3 columns)
#>    1. spc: I[fl]/"a.u." [matrix, array181] 27.15000 66.80133 ... 294.6495 
#>    2. filename: filename [character] rawdata/flu1.txt rawdata/flu2.txt ... rawdata/flu6.txt 
#>    3. c: c / (mg / l) [numeric] 0.05 0.10 ... 0.3