Load and attach hyperSpec and all installed R hyperSpec family (hySpc.*) packages

This function loads and attaches hyperSpec and all installed R hyperSpec family packages, which share common hySpc.something package name pattern.

hy_attach(exclude = "hySpc.testthat", ..., quiet = NA)

Arguments

exclude
Character vector with the names of the packages that should not be explicitly loaded. If they are already loaded and/or attached, they will not be unloaded or detached.

...
Further parameters to base::library().

quiet
(TRUE|NA|FALSE) Controls the verbosity of messages:

FALSE: no messages will be printed to the console.

NA (default): short message with the names of hySpc.* packages to load and attach will be printed, but package startup messages will be suppressed.

TRUE: both short and package startup messages will be printed.

Value

Invisibly returns a list with the output of base::library() for each R hyperSpec family package that was loaded and attached.

See also

Author

V. Gegzna and C. Beleites

Examples

if (FALSE) {
hyperSpec::hy_attach()

hyperSpec::hy_attach(quiet = TRUE)
}