List available R hyperSpec family packages

Get the names of R hyperSpec family packages that are available on GitHub organization r-hyperspec (the official homepage of these packages).

This function requires Internet connection.

hy_list_available_hySpc_packages()

Value

Character vector with the names of the pacakges.

Details

If your machine is connected to the Internet and you receive issues due to overused limit of anonymous connections to GitHub, you may run alternative code that employs GitHub PAT (personal access token) for authentication. On how to set your GitHub PAT, see sections "Create a PAT" and "Add your PAT to .Renviron" in this book: happygitwithr.com/github-pat.html.

# install.pacakges("gh")
gh_api_response <- gh::gh("GET /users/r-hyperspec/repos?per_page=100")
repo_names <- vapply(gh_api_response, "[[", "", "name")
package_names <- grep("^hyperSpec|^hySpc[.]", repo_names, value = TRUE)
package_names

Author

V. Gegzna

Examples