Class hyperSpec

This package gives an interface to handle hyperspectral data sets in R. Hyperspectral data are spatially or time-resolved spectra, or spectra with any other kind of information associated with the spectra. E.g. spectral maps or images, time series, calibration series, etc.
The spectra can be data as obtained in XRF, UV/VIS, Fluorescence, AES, NIR, IR, Raman, NMR, MS, etc.
More generally, any data that is recorded over a discretized variable, e.g. absorbance = f(wavelength), stored as a vector of absorbance values for discrete wavelengths is suitable.

Slots

wavelength

wavelengths (wavenumbers, frequencies, etc.) for each of the columns of the spectra matrix

data

the data (extra data and spectra matrix)

label

expressions for column labels (incl. units). The label of the wavelength axis is in the special element .wavelength.

log

deprecated.

Note

Please note that the logbook is now removed.

See also

Author

C. Beleites

Examples

showClass("hyperSpec")

#> Class "hyperSpec" [package "hyperSpec"]
#> 
#> Slots:
#>                                                   
#> Name:  wavelength       data      label        log
#> Class:    numeric data.frame       list data.frame

if (FALSE) {
vignette("hyperSpec")
}