Math Functions for hyperSpec Objects

Mathematical functions for hyperSpec Objects.

# S4 method for hyperSpec
Math2(x, digits)

# S4 method for hyperSpec
log(x, base = exp(1), ...)

# S4 method for hyperSpec
Math(x)

Arguments

x
the hyperSpec object

digits
integer stating the rounding precision

base
base of logarithm

...
ignored

Value

a hyperSpec object

Details

The functions abs, sign, sqrt, floor, ceiling, trunc, round, signif, exp, log, expm1, log1p, cos, sin, tan, acos, asin, atan, cosh, sinh, tanh, acosh, asinh, atanh, lgamma, gamma, digamma, trigamma, cumsum, cumprod, cummax, cummin for hyperSpec objects.

See also

Author

C. Beleites

Examples

	log (fllogflu

#> hyperSpec object
#>    6 spectra
#>    3 data columns
#>    181 data points / spectrum
#> wavelength: lambda/nm [numeric] 405.0 405.5 ... 495 
#> data:  (6 rows x 3 columns)
#>    1. spc: I[fl]/"a.u." [matrix, array181] 3.301377 4.201723 ... 5.685787 
#>    2. filename: filename [character] rawdata/flu1.txt rawdata/flu2.txt ... rawdata/flu6.txt 
#>    3. c: c / (mg / l) [numeric] 0.05 0.10 ... 0.3