Merge hyperSpec objects

Merges two hyperSpec objects and cbinds their spectra matrices, or merges additional extra data into a hyperSpec object.

# S4 method for hyperSpec,hyperSpec
merge(x, y, ...)

# S4 method for hyperSpec,data.frame
merge(x, y, ...)

# S4 method for data.frame,hyperSpec
merge(x, y, ...)

Arguments

x
a hyperSpec object or data.frame

y
a hyperSpec object or data.frame (including derived classes like tibble)

...
handed to merge.data.frame

Details

After merging, the spectra matrix can contain duplicates, and is not ordered according to the wavelength.

If the wavelength axis should be ordered, use orderwl.

If a hyperSpec object and a data.frame are merged, the result is of the class of the first (x) object.

See also

Author

C. Beleites

Examples

merge (chondro [1:10,, 600], chondro [5:15,, 600], by = c("x", "y"))$.

#>       x     y          filename.x clusters.x .nx          filename.y clusters.y
#> 1 -2.55 -4.77 rawdata/chondro.txt     lacuna  10 rawdata/chondro.txt     lacuna
#> 2 -3.55 -4.77 rawdata/chondro.txt     lacuna   9 rawdata/chondro.txt     lacuna
#> 3 -4.55 -4.77 rawdata/chondro.txt     lacuna   8 rawdata/chondro.txt     lacuna
#> 4 -5.55 -4.77 rawdata/chondro.txt     lacuna   7 rawdata/chondro.txt     lacuna
#> 5 -6.55 -4.77 rawdata/chondro.txt     matrix   6 rawdata/chondro.txt     matrix
#> 6 -7.55 -4.77 rawdata/chondro.txt     matrix   5 rawdata/chondro.txt     matrix
#>   .ny spc.1 spc.2
#> 1   6    NA    NA
#> 2   5    NA    NA
#> 3   4    NA    NA
#> 4   3    NA    NA
#> 5   2    NA    NA
#> 6   1    NA    NA

tmp <- merge (chondro [1:10,, 610], chondro [5:15,, 610],
              by = c("x", "y"), all = TRUE)
tmp$.

#>         x     y          filename.x clusters.x .nx          filename.y
#> 1  -11.55 -4.77 rawdata/chondro.txt     matrix   1                <NA>
#> 2  -10.55 -4.77 rawdata/chondro.txt     matrix   2                <NA>
#> 3   -9.55 -4.77 rawdata/chondro.txt     matrix   3                <NA>
#> 4   -8.55 -4.77 rawdata/chondro.txt     matrix   4                <NA>
#> 5   -7.55 -4.77 rawdata/chondro.txt     matrix   5 rawdata/chondro.txt
#> 6   -6.55 -4.77 rawdata/chondro.txt     matrix   6 rawdata/chondro.txt
#> 7   -5.55 -4.77 rawdata/chondro.txt     lacuna   7 rawdata/chondro.txt
#> 8   -4.55 -4.77 rawdata/chondro.txt     lacuna   8 rawdata/chondro.txt
#> 9   -3.55 -4.77 rawdata/chondro.txt     lacuna   9 rawdata/chondro.txt
#> 10  -2.55 -4.77 rawdata/chondro.txt     lacuna  10 rawdata/chondro.txt
#> 11  -1.55 -4.77                <NA>       <NA>  NA rawdata/chondro.txt
#> 12  -0.55 -4.77                <NA>       <NA>  NA rawdata/chondro.txt
#> 13   0.45 -4.77                <NA>       <NA>  NA rawdata/chondro.txt
#> 14   1.45 -4.77                <NA>       <NA>  NA rawdata/chondro.txt
#> 15   2.45 -4.77                <NA>       <NA>  NA rawdata/chondro.txt
#>    clusters.y .ny    spc.1    spc.2
#> 1        <NA>  NA 488.6323       NA
#> 2        <NA>  NA 489.4758       NA
#> 3        <NA>  NA 456.0324       NA
#> 4        <NA>  NA 464.8207       NA
#> 5      matrix   1 428.6620 428.6620
#> 6      matrix   2 426.0734 426.0734
#> 7      lacuna   3 412.3674 412.3674
#> 8      lacuna   4 398.4772 398.4772
#> 9      lacuna   5 412.1881 412.1881
#> 10     lacuna   6 392.5067 392.5067
#> 11     lacuna   7       NA 376.6938
#> 12     lacuna   8       NA 357.0593
#> 13     lacuna   9       NA 368.5321
#> 14     lacuna  10       NA 357.8674
#> 15     lacuna  11       NA 365.9732

wl (tmp)

#> [1] 610 610

## remove duplicated wavelengths:
approxfun <- function (y, wl, new.wl){
  approx (wl, y, new.wl, method = "constant",
          ties = function (x) mean (x, na.rm = TRUE)
          )$y
}

merged <- merge (chondro [1:7,, 610 ~ 620], chondro [5:10,, 615 ~ 625], all = TRUE)
merged$.

#>        y      x            filename clusters .nx .ny    spc.1    spc.2    spc.3
#> 1  -4.77 -11.55 rawdata/chondro.txt   matrix   1  NA 488.6323 466.1774 492.0016
#> 2  -4.77 -10.55 rawdata/chondro.txt   matrix   2  NA 489.4758 465.0506 490.5329
#> 3  -4.77  -9.55 rawdata/chondro.txt   matrix   3  NA 456.0324 436.6220 458.0576
#> 4  -4.77  -8.55 rawdata/chondro.txt   matrix   4  NA 464.8207 444.8486 470.0172
#> 5  -4.77  -7.55 rawdata/chondro.txt   matrix   5   1 428.6620 410.7956 433.1228
#> 6  -4.77  -6.55 rawdata/chondro.txt   matrix   6   2 426.0734 407.8570 431.2145
#> 7  -4.77  -5.55 rawdata/chondro.txt   lacuna   7   3 412.3674 396.5000 421.2738
#> 8  -4.77  -4.55 rawdata/chondro.txt   lacuna  NA   4       NA       NA       NA
#> 9  -4.77  -3.55 rawdata/chondro.txt   lacuna  NA   5       NA       NA       NA
#> 10 -4.77  -2.55 rawdata/chondro.txt   lacuna  NA   6       NA       NA       NA
#>       spc.4    spc.5    spc.6    spc.7
#> 1        NA       NA       NA       NA
#> 2        NA       NA       NA       NA
#> 3        NA       NA       NA       NA
#> 4        NA       NA       NA       NA
#> 5  410.7956 433.1228 461.1903 397.3773
#> 6  407.8570 431.2145 458.1503 394.1776
#> 7  396.5000 421.2738 445.5431 382.7198
#> 8  381.9505 406.2471 429.6729 368.4599
#> 9  397.5075 423.3002 446.1478 381.8675
#> 10 377.3918 402.2349 424.1902 362.4297

merged <- apply (merged, 1, approxfun,
                 wl = wl (merged), new.wl = unique (wl (merged)),
                 new.wavelength = "new.wl")
merged$.

#>        y      x            filename clusters .nx .ny    spc.1    spc.2    spc.3
#> 1  -4.77 -11.55 rawdata/chondro.txt   matrix   1  NA 488.6323 466.1774 492.0016
#> 2  -4.77 -10.55 rawdata/chondro.txt   matrix   2  NA 489.4758 465.0506 490.5329
#> 3  -4.77  -9.55 rawdata/chondro.txt   matrix   3  NA 456.0324 436.6220 458.0576
#> 4  -4.77  -8.55 rawdata/chondro.txt   matrix   4  NA 464.8207 444.8486 470.0172
#> 5  -4.77  -7.55 rawdata/chondro.txt   matrix   5   1 428.6620 410.7956 433.1228
#> 6  -4.77  -6.55 rawdata/chondro.txt   matrix   6   2 426.0734 407.8570 431.2145
#> 7  -4.77  -5.55 rawdata/chondro.txt   lacuna   7   3 412.3674 396.5000 421.2738
#> 8  -4.77  -4.55 rawdata/chondro.txt   lacuna  NA   4       NA 381.9505 406.2471
#> 9  -4.77  -3.55 rawdata/chondro.txt   lacuna  NA   5       NA 397.5075 423.3002
#> 10 -4.77  -2.55 rawdata/chondro.txt   lacuna  NA   6       NA 377.3918 402.2349
#>       spc.4    spc.5
#> 1        NA       NA
#> 2        NA       NA
#> 3        NA       NA
#> 4        NA       NA
#> 5  461.1903 397.3773
#> 6  458.1503 394.1776
#> 7  445.5431 382.7198
#> 8  429.6729 368.4599
#> 9  446.1478 381.8675
#> 10 424.1902 362.4297

## merging data.frame into hyperSpec object => hyperSpec object
y <- data.frame (filename = sample (flu$filename, 4, replace = TRUE), cpred = 1:4)
y

#>           filename cpred
#> 1 rawdata/flu1.txt     1
#> 2 rawdata/flu6.txt     2
#> 3 rawdata/flu1.txt     3
#> 4 rawdata/flu6.txt     4

tmp <- merge (flu, y)
tmp$..

#>           filename    c cpred
#> 1 rawdata/flu1.txt 0.05     1
#> 2 rawdata/flu1.txt 0.05     3
#> 3 rawdata/flu6.txt 0.30     2
#> 4 rawdata/flu6.txt 0.30     4

## merging hyperSpec object into data.frame => data.frame
merge (y, flu)

#>           filename cpred   spc.405 spc.405.5   spc.406 spc.406.5   spc.407
#> 1 rawdata/flu1.txt     1  27.15000  32.34467  33.37867  34.41933  36.53133
#> 2 rawdata/flu1.txt     3  27.15000  32.34467  33.37867  34.41933  36.53133
#> 3 rawdata/flu6.txt     2 198.43033 209.45800 215.78500 224.58700 232.52800
#> 4 rawdata/flu6.txt     4 198.43033 209.45800 215.78500 224.58700 232.52800
#>   spc.407.5   spc.408 spc.408.5   spc.409 spc.409.5   spc.410 spc.410.5
#> 1  37.64767  38.13700  39.17700  40.73567  41.38133  44.25133  44.12633
#> 2  37.64767  38.13700  39.17700  40.73567  41.38133  44.25133  44.12633
#> 3 240.77133 248.04667 256.89133 262.73900 270.27133 281.82867 287.70467
#> 4 240.77133 248.04667 256.89133 262.73900 270.27133 281.82867 287.70467
#>     spc.411 spc.411.5   spc.412 spc.412.5   spc.413 spc.413.5   spc.414
#> 1  46.98067  49.08167  50.27433  50.10967  52.23167  53.03967  54.51900
#> 2  46.98067  49.08167  50.27433  50.10967  52.23167  53.03967  54.51900
#> 3 297.23500 307.59333 314.99367 324.56300 330.80033 340.89667 347.62167
#> 4 297.23500 307.59333 314.99367 324.56300 330.80033 340.89667 347.62167
#>   spc.414.5   spc.415 spc.415.5   spc.416 spc.416.5   spc.417 spc.417.5
#> 1  56.22033  57.71933  59.51433  58.74500  60.09467  61.84067  62.16867
#> 2  56.22033  57.71933  59.51433  58.74500  60.09467  61.84067  62.16867
#> 3 355.02667 365.62333 374.35700 380.21733 391.40100 398.17300 405.42933
#> 4 355.02667 365.62333 374.35700 380.21733 391.40100 398.17300 405.42933
#>     spc.418 spc.418.5   spc.419 spc.419.5   spc.420 spc.420.5   spc.421
#> 1  66.04933  65.24500  65.79900  66.94333  69.13600  70.92467  73.83400
#> 2  66.04933  65.24500  65.79900  66.94333  69.13600  70.92467  73.83400
#> 3 412.48400 421.27300 429.36267 433.35767 443.18733 447.88033 461.33933
#> 4 412.48400 421.27300 429.36267 433.35767 443.18733 447.88033 461.33933
#>   spc.421.5   spc.422 spc.422.5   spc.423 spc.423.5   spc.424 spc.424.5
#> 1  74.38567  74.17333  75.53067  76.42967  76.13200  77.89067  79.36867
#> 2  74.38567  74.17333  75.53067  76.42967  76.13200  77.89067  79.36867
#> 3 467.12733 475.18967 481.66667 488.73567 494.01867 504.47333 515.40467
#> 4 467.12733 475.18967 481.66667 488.73567 494.01867 504.47333 515.40467
#>     spc.425 spc.425.5   spc.426 spc.426.5   spc.427 spc.427.5   spc.428
#> 1  79.42733  82.02833  83.87833  83.81367  85.50667  86.50233  88.93733
#> 2  79.42733  82.02833  83.87833  83.81367  85.50667  86.50233  88.93733
#> 3 521.23867 522.64633 536.03533 544.41767 551.63333 556.71267 567.61267
#> 4 521.23867 522.64633 536.03533 544.41767 551.63333 556.71267 567.61267
#>   spc.428.5   spc.429 spc.429.5   spc.430 spc.430.5   spc.431 spc.431.5
#> 1  88.99533  89.51500  90.55667  91.70633  93.57900  94.01300  94.21033
#> 2  88.99533  89.51500  90.55667  91.70633  93.57900  94.01300  94.21033
#> 3 572.54933 578.72667 587.74900 590.54867 600.42300 608.35800 614.03833
#> 4 572.54933 578.72667 587.74900 590.54867 600.42300 608.35800 614.03833
#>     spc.432 spc.432.5   spc.433 spc.433.5   spc.434 spc.434.5   spc.435
#> 1  96.44200  96.62733  98.96233  98.84767 100.56833 100.44933  99.91600
#> 2  96.44200  96.62733  98.96233  98.84767 100.56833 100.44933  99.91600
#> 3 616.95567 621.87267 629.80033 633.44500 637.13233 639.16167 642.89833
#> 4 616.95567 621.87267 629.80033 633.44500 637.13233 639.16167 642.89833
#>   spc.435.5   spc.436 spc.436.5   spc.437 spc.437.5   spc.438 spc.438.5
#> 1  99.71900 100.88633 101.44533 102.64267 104.22967 103.11133 104.75800
#> 2  99.71900 100.88633 101.44533 102.64267 104.22967 103.11133 104.75800
#> 3 648.17200 651.67900 654.53967 656.83267 659.77800 662.05633 663.12300
#> 4 648.17200 651.67900 654.53967 656.83267 659.77800 662.05633 663.12300
#>     spc.439 spc.439.5   spc.440 spc.440.5   spc.441 spc.441.5   spc.442
#> 1 103.15833 103.94267 105.37100 105.86333 103.67967 104.09000 104.65400
#> 2 103.15833 103.94267 105.37100 105.86333 103.67967 104.09000 104.65400
#> 3 663.48200 670.47567 669.88700 670.06033 671.00800 667.61233 666.30833
#> 4 663.48200 670.47567 669.88700 670.06033 671.00800 667.61233 666.30833
#>   spc.442.5   spc.443 spc.443.5   spc.444 spc.444.5   spc.445 spc.445.5
#> 1 105.22633 104.80433 105.81033 104.75800 103.88300 105.36167 105.01500
#> 2 105.22633 104.80433 105.81033 104.75800 103.88300 105.36167 105.01500
#> 3 669.97433 675.68300 675.97867 676.64200 674.73667 671.23433 673.61133
#> 4 669.97433 675.68300 675.97867 676.64200 674.73667 671.23433 673.61133
#>     spc.446 spc.446.5   spc.447 spc.447.5   spc.448 spc.448.5   spc.449
#> 1 105.05933 105.34933 103.98700 105.43867 104.19667 105.09800 104.72333
#> 2 105.05933 105.34933 103.98700 105.43867 104.19667 105.09800 104.72333
#> 3 676.73533 677.49467 674.28033 676.64567 674.89067 673.09200 674.89800
#> 4 676.73533 677.49467 674.28033 676.64567 674.89067 673.09200 674.89800
#>   spc.449.5   spc.450 spc.450.5   spc.451 spc.451.5   spc.452 spc.452.5
#> 1 106.66733 106.94967 104.75467 105.08267 105.30000 105.21300 104.78100
#> 2 106.66733 106.94967 104.75467 105.08267 105.30000 105.21300 104.78100
#> 3 671.64167 672.53167 672.28033 671.00033 672.87900 674.74300 672.72467
#> 4 671.64167 672.53167 672.28033 671.00033 672.87900 674.74300 672.72467
#>     spc.453 spc.453.5   spc.454 spc.454.5   spc.455 spc.455.5   spc.456
#> 1 104.53867 105.13267 105.16967 104.05700 106.38533 104.07967 104.40133
#> 2 104.53867 105.13267 105.16967 104.05700 106.38533 104.07967 104.40133
#> 3 669.79100 672.00433 672.33967 668.74033 662.63333 666.29700 664.63267
#> 4 669.79100 672.00433 672.33967 668.74033 662.63333 666.29700 664.63267
#>   spc.456.5   spc.457 spc.457.5   spc.458 spc.458.5   spc.459 spc.459.5
#> 1 102.18133 103.44167 101.79700 102.87167 102.38867 100.41900 101.16167
#> 2 102.18133 103.44167 101.79700 102.87167 102.38867 100.41900 101.16167
#> 3 659.91800 660.13033 656.23733 654.19500 652.31833 647.82100 643.38133
#> 4 659.91800 660.13033 656.23733 654.19500 652.31833 647.82100 643.38133
#>     spc.460 spc.460.5   spc.461 spc.461.5   spc.462 spc.462.5   spc.463
#> 1  98.61133  98.42933  98.57600  98.34100  98.46733  95.14900  94.71100
#> 2  98.61133  98.42933  98.57600  98.34100  98.46733  95.14900  94.71100
#> 3 640.79033 632.11967 633.56367 627.81900 623.21133 619.99667 613.55133
#> 4 640.79033 632.11967 633.56367 627.81900 623.21133 619.99667 613.55133
#>   spc.463.5   spc.464 spc.464.5   spc.465 spc.465.5   spc.466 spc.466.5
#> 1  95.27433  94.88433  93.62133  92.46567  92.11533  91.30867  89.53900
#> 2  95.27433  94.88433  93.62133  92.46567  92.11533  91.30867  89.53900
#> 3 606.01267 605.92100 599.88400 593.00650 587.40150 586.38900 581.49850
#> 4 606.01267 605.92100 599.88400 593.00650 587.40150 586.38900 581.49850
#>     spc.467 spc.467.5   spc.468 spc.468.5   spc.469 spc.469.5   spc.470
#> 1  88.28100  86.13633  86.16200  86.81733  85.95767  84.06700  86.21200
#> 2  88.28100  86.13633  86.16200  86.81733  85.95767  84.06700  86.21200
#> 3 574.15100 566.49250 560.29000 559.46250 550.27850 545.54350 542.76900
#> 4 574.15100 566.49250 560.29000 559.46250 550.27850 545.54350 542.76900
#>   spc.470.5   spc.471 spc.471.5   spc.472 spc.472.5   spc.473 spc.473.5
#> 1  83.68333  82.56867  82.70533  81.26033  78.75233  76.67533  77.95300
#> 2  83.68333  82.56867  82.70533  81.26033  78.75233  76.67533  77.95300
#> 3 528.92800 528.12900 519.65350 516.54950 512.45650 502.48600 499.65250
#> 4 528.92800 528.12900 519.65350 516.54950 512.45650 502.48600 499.65250
#>     spc.474 spc.474.5   spc.475 spc.475.5   spc.476 spc.476.5   spc.477
#> 1  77.50700  76.16400  75.64600  76.43700  74.57033  72.87900  72.80267
#> 2  77.50700  76.16400  75.64600  76.43700  74.57033  72.87900  72.80267
#> 3 495.00100 486.49700 481.73950 476.47200 472.37900 469.25450 466.99550
#> 4 495.00100 486.49700 481.73950 476.47200 472.37900 469.25450 466.99550
#>   spc.477.5   spc.478 spc.478.5   spc.479 spc.479.5   spc.480 spc.480.5
#> 1  71.05200  70.21367  69.60533  69.98967  68.78000  68.34233  67.68467
#> 2  71.05200  70.21367  69.60533  69.98967  68.78000  68.34233  67.68467
#> 3 457.76300 452.32300 450.48350 445.14500 444.27400 435.22800 430.33300
#> 4 457.76300 452.32300 450.48350 445.14500 444.27400 435.22800 430.33300
#>     spc.481 spc.481.5   spc.482 spc.482.5   spc.483 spc.483.5   spc.484
#> 1  67.27700  67.04833  65.31300  64.50900  63.50567  62.18367  62.04500
#> 2  67.27700  67.04833  65.31300  64.50900  63.50567  62.18367  62.04500
#> 3 424.72000 420.74800 418.46650 414.01750 408.32400 403.85400 405.00450
#> 4 424.72000 420.74800 418.46650 414.01750 408.32400 403.85400 405.00450
#>   spc.484.5   spc.485 spc.485.5   spc.486 spc.486.5   spc.487 spc.487.5
#> 1  62.02733  61.79900  60.52800  59.34200  59.12500  57.74800  57.44667
#> 2  62.02733  61.79900  60.52800  59.34200  59.12500  57.74800  57.44667
#> 3 394.33500 389.97750 384.49100 382.33700 375.08200 369.09900 366.06300
#> 4 394.33500 389.97750 384.49100 382.33700 375.08200 369.09900 366.06300
#>     spc.488 spc.488.5   spc.489 spc.489.5   spc.490 spc.490.5   spc.491
#> 1  57.74267  56.27500  55.49233  54.40933  53.83267  53.26333  52.45733
#> 2  57.74267  56.27500  55.49233  54.40933  53.83267  53.26333  52.45733
#> 3 362.45350 358.99650 352.14300 348.20000 347.10650 338.79150 335.34700
#> 4 362.45350 358.99650 352.14300 348.20000 347.10650 338.79150 335.34700
#>   spc.491.5   spc.492 spc.492.5   spc.493 spc.493.5   spc.494 spc.494.5
#> 1  52.14000  49.78433  49.62300  48.33800  47.30400  47.16267  46.41233
#> 2  52.14000  49.78433  49.62300  48.33800  47.30400  47.16267  46.41233
#> 3 330.04000 325.16200 322.06850 315.64550 312.32950 307.51850 302.32550
#> 4 330.04000 325.16200 322.06850 315.64550 312.32950 307.51850 302.32550
#>     spc.495    c
#> 1  45.25633 0.05
#> 2  45.25633 0.05
#> 3 294.64950 0.30
#> 4 294.64950 0.30