File Import Witec Raman

Import Raman Spectra/Maps from Witec Instrument via ASCII files

scan.txt.Witec(...)

scan.dat.Witec(...)

scan.txt.Witec.Graph(...)

read.txt.Witec(
  file = stop("filename or connection needed"),
  points.per.line = NULL,
  lines.per.image = NULL,
  type = c("single", "map"),
  hdr.label = FALSE,
  hdr.units = FALSE,
  encoding = "unknown",
  ...,
  quiet = TRUE
)

read.dat.Witec(
  filex = stop("filename or connection needed"),
  filey = sub("-x", "-y", filex),
  points.per.line = NULL,
  lines.per.image = NULL,
  type = c("single", "map"),
  encoding = "unknown",
  ...,
  quiet = hy.getOption("debuglevel") < 1L
)

read.txt.Witec.Graph(
  headerfile = stop("filename or connection needed"),
  filex = gsub("Header", "X-Axis", headerfile),
  filey = gsub("Header", "Y-Axis", headerfile),
  type = c("single", "map"),
  encoding = "unknown",
  ...,
  quiet = TRUE
)

Arguments

..., quiet
handed to scan

file
filename or connection to ASCII file

points.per.line
number of spectra in x direction of the map

lines.per.image
number of spectra in y direction

type
type of spectra: single for single spectra (including time series), map for imaging data.

hdr.label
WITec Project exports the spectra names (contain information of map position or number of spectra) within the file.

hdr.units
WITec Project exports the spectra units within the file.

encoding
character encoding, see readLines

filex
filename wavelength axis file

filey
filename intensity file

headerfile
filename or connection to ASCII file with header information

nwl
is deprecated and will be removed soon. Number of wavelengths is calculated automatically.

remove.zerospc
is deprecated and will be removed soon. Use hy.setOptions (file.remove.emptyspc = TRUE) instead.

Value

a hyperSpec object

Details

read.txt.Witec reads Witec ASCII files where the first column gives the wavelength axes and the other columns the spectra. read.dat.Witec reads Witec's ASCII exported data which comes in separate files with x and y data.

See also

Author

Claudia Beleites and Marcel Dahms