Impute missing data points

Replace NAs in the spectra matrix by interpolation. With less than 4 points available linear interpolation of the 2 neighbour points is used. For larger numbers of neighbour points, smoothing interpolation is performed by smooth.spline.

spc.NA.approx(
  spc,
  neighbours = 1,
  w = rep(1, 2 * neighbours),
  df = 1 + .Machine$double.eps,
  spar = NULL,
  debuglevel = hy.getOption("debuglevel")
)

spc.NA.linapprox(...)

Arguments

spc
hyperSpec object with spectra matrix containing NAs

neighbours
how many neighbour data points should be used to fit the line

w, df, spar
see smooth.spline

debuglevel
see options

...
ignored

Value

hyperSpec object

Note

The function has been renamed from spc.NA.linapprox to spc.NA.approx

Author

Claudia Beleites

Examples

fluNA <- hyperSpec:::fluNA

#> Error in get(name, envir = asNamespace(pkg), inherits = FALSE): object 'fluNA' not found

spc.NA.approx (fluNA [,, min ~ 410], debuglevel = 1)

#> Error in is(object, "hyperSpec"): object 'fluNA' not found

spc.NA.approx (fluNA [1,, min ~ 410], debuglevel = 2)

#> Error in is(object, "hyperSpec"): object 'fluNA' not found

spc.NA.approx (fluNA [4,, min ~ 410], neighbours = 3, df = 4, debuglevel = 2)

#> Error in is(object, "hyperSpec"): object 'fluNA' not found