Merge hyperSpec objects

Merges two hyperSpec objects and cbinds their spectra matrices, or merges additional extra data into a hyperSpec object.

# S4 method for hyperSpec,hyperSpec
merge(x, y, ...)

# S4 method for hyperSpec,data.frame
merge(x, y, ...)

# S4 method for data.frame,hyperSpec
merge(x, y, ...)

Arguments

x
a hyperSpec object or data.frame

y
a hyperSpec object or data.frame (including derived classes like tibble)

...
handed to merge.data.frame

Details

After merging, the spectra matrix can contain duplicates, and is not ordered according to the wavelength.

If the wavelength axis should be ordered, use orderwl.

If a hyperSpec object and a data.frame are merged, the result is of the class of the first (x) object.

See also

Author

C. Beleites

Examples

merge (chondro [1:10,, 600], chondro [5:15,, 600], by = c("x", "y"))$.

#>       x     y          filename.x clusters.x .nx          filename.y clusters.y
#> 1 -2.55 -4.77 rawdata/chondro.txt     lacuna  10 rawdata/chondro.txt     lacuna
#> 2 -3.55 -4.77 rawdata/chondro.txt     lacuna   9 rawdata/chondro.txt     lacuna
#> 3 -4.55 -4.77 rawdata/chondro.txt     lacuna   8 rawdata/chondro.txt     lacuna
#> 4 -5.55 -4.77 rawdata/chondro.txt     lacuna   7 rawdata/chondro.txt     lacuna
#> 5 -6.55 -4.77 rawdata/chondro.txt     matrix   6 rawdata/chondro.txt     matrix
#> 6 -7.55 -4.77 rawdata/chondro.txt     matrix   5 rawdata/chondro.txt     matrix
#>   .ny spc.1 spc.2
#> 1   6    NA    NA
#> 2   5    NA    NA
#> 3   4    NA    NA
#> 4   3    NA    NA
#> 5   2    NA    NA
#> 6   1    NA    NA

tmp <- merge (chondro [1:10,, 610], chondro [5:15,, 610],
              by = c("x", "y"), all = TRUE)
tmp$.

#>         x     y          filename.x clusters.x .nx          filename.y
#> 1  -11.55 -4.77 rawdata/chondro.txt     matrix   1                <NA>
#> 2  -10.55 -4.77 rawdata/chondro.txt     matrix   2                <NA>
#> 3   -9.55 -4.77 rawdata/chondro.txt     matrix   3                <NA>
#> 4   -8.55 -4.77 rawdata/chondro.txt     matrix   4                <NA>
#> 5   -7.55 -4.77 rawdata/chondro.txt     matrix   5 rawdata/chondro.txt
#> 6   -6.55 -4.77 rawdata/chondro.txt     matrix   6 rawdata/chondro.txt
#> 7   -5.55 -4.77 rawdata/chondro.txt     lacuna   7 rawdata/chondro.txt
#> 8   -4.55 -4.77 rawdata/chondro.txt     lacuna   8 rawdata/chondro.txt
#> 9   -3.55 -4.77 rawdata/chondro.txt     lacuna   9 rawdata/chondro.txt
#> 10  -2.55 -4.77 rawdata/chondro.txt     lacuna  10 rawdata/chondro.txt
#> 11  -1.55 -4.77                <NA>       <NA>  NA rawdata/chondro.txt
#> 12  -0.55 -4.77                <NA>       <NA>  NA rawdata/chondro.txt
#> 13   0.45 -4.77                <NA>       <NA>  NA rawdata/chondro.txt
#> 14   1.45 -4.77                <NA>       <NA>  NA rawdata/chondro.txt
#> 15   2.45 -4.77                <NA>       <NA>  NA rawdata/chondro.txt
#>    clusters.y .ny    spc.1    spc.2
#> 1        <NA>  NA 488.6323       NA
#> 2        <NA>  NA 489.4758       NA
#> 3        <NA>  NA 456.0324       NA
#> 4        <NA>  NA 464.8207       NA
#> 5      matrix   1 428.6620 428.6620
#> 6      matrix   2 426.0734 426.0734
#> 7      lacuna   3 412.3674 412.3674
#> 8      lacuna   4 398.4772 398.4772
#> 9      lacuna   5 412.1881 412.1881
#> 10     lacuna   6 392.5067 392.5067
#> 11     lacuna   7       NA 376.6938
#> 12     lacuna   8       NA 357.0593
#> 13     lacuna   9       NA 368.5321
#> 14     lacuna  10       NA 357.8674
#> 15     lacuna  11       NA 365.9732

wl (tmp)

#> [1] 610 610

## remove duplicated wavelengths:
approxfun <- function (y, wl, new.wl){
  approx (wl, y, new.wl, method = "constant",
          ties = function (x) mean (x, na.rm = TRUE)
          )$y
}

merged <- merge (chondro [1:7,, 610 ~ 620], chondro [5:10,, 615 ~ 625], all = TRUE)
merged$.

#>        y      x            filename clusters .nx .ny    spc.1    spc.2    spc.3
#> 1  -4.77 -11.55 rawdata/chondro.txt   matrix   1  NA 488.6323 466.1774 492.0016
#> 2  -4.77 -10.55 rawdata/chondro.txt   matrix   2  NA 489.4758 465.0506 490.5329
#> 3  -4.77  -9.55 rawdata/chondro.txt   matrix   3  NA 456.0324 436.6220 458.0576
#> 4  -4.77  -8.55 rawdata/chondro.txt   matrix   4  NA 464.8207 444.8486 470.0172
#> 5  -4.77  -7.55 rawdata/chondro.txt   matrix   5   1 428.6620 410.7956 433.1228
#> 6  -4.77  -6.55 rawdata/chondro.txt   matrix   6   2 426.0734 407.8570 431.2145
#> 7  -4.77  -5.55 rawdata/chondro.txt   lacuna   7   3 412.3674 396.5000 421.2738
#> 8  -4.77  -4.55 rawdata/chondro.txt   lacuna  NA   4       NA       NA       NA
#> 9  -4.77  -3.55 rawdata/chondro.txt   lacuna  NA   5       NA       NA       NA
#> 10 -4.77  -2.55 rawdata/chondro.txt   lacuna  NA   6       NA       NA       NA
#>       spc.4    spc.5    spc.6    spc.7
#> 1        NA       NA       NA       NA
#> 2        NA       NA       NA       NA
#> 3        NA       NA       NA       NA
#> 4        NA       NA       NA       NA
#> 5  410.7956 433.1228 461.1903 397.3773
#> 6  407.8570 431.2145 458.1503 394.1776
#> 7  396.5000 421.2738 445.5431 382.7198
#> 8  381.9505 406.2471 429.6729 368.4599
#> 9  397.5075 423.3002 446.1478 381.8675
#> 10 377.3918 402.2349 424.1902 362.4297

merged <- apply (merged, 1, approxfun,
                 wl = wl (merged), new.wl = unique (wl (merged)),
                 new.wavelength = "new.wl")
merged$.

#>        y      x            filename clusters .nx .ny    spc.1    spc.2    spc.3
#> 1  -4.77 -11.55 rawdata/chondro.txt   matrix   1  NA 488.6323 466.1774 492.0016
#> 2  -4.77 -10.55 rawdata/chondro.txt   matrix   2  NA 489.4758 465.0506 490.5329
#> 3  -4.77  -9.55 rawdata/chondro.txt   matrix   3  NA 456.0324 436.6220 458.0576
#> 4  -4.77  -8.55 rawdata/chondro.txt   matrix   4  NA 464.8207 444.8486 470.0172
#> 5  -4.77  -7.55 rawdata/chondro.txt   matrix   5   1 428.6620 410.7956 433.1228
#> 6  -4.77  -6.55 rawdata/chondro.txt   matrix   6   2 426.0734 407.8570 431.2145
#> 7  -4.77  -5.55 rawdata/chondro.txt   lacuna   7   3 412.3674 396.5000 421.2738
#> 8  -4.77  -4.55 rawdata/chondro.txt   lacuna  NA   4       NA 381.9505 406.2471
#> 9  -4.77  -3.55 rawdata/chondro.txt   lacuna  NA   5       NA 397.5075 423.3002
#> 10 -4.77  -2.55 rawdata/chondro.txt   lacuna  NA   6       NA 377.3918 402.2349
#>       spc.4    spc.5
#> 1        NA       NA
#> 2        NA       NA
#> 3        NA       NA
#> 4        NA       NA
#> 5  461.1903 397.3773
#> 6  458.1503 394.1776
#> 7  445.5431 382.7198
#> 8  429.6729 368.4599
#> 9  446.1478 381.8675
#> 10 424.1902 362.4297

## merging data.frame into hyperSpec object => hyperSpec object
y <- data.frame (filename = sample (flu$filename, 4, replace = TRUE), cpred = 1:4)
y

#>           filename cpred
#> 1 rawdata/flu2.txt     1
#> 2 rawdata/flu6.txt     2
#> 3 rawdata/flu5.txt     3
#> 4 rawdata/flu6.txt     4

tmp <- merge (flu, y)
tmp$..

#>           filename    c cpred
#> 1 rawdata/flu2.txt 0.10     1
#> 2 rawdata/flu5.txt 0.25     3
#> 3 rawdata/flu6.txt 0.30     2
#> 4 rawdata/flu6.txt 0.30     4

## merging hyperSpec object into data.frame => data.frame
merge (y, flu)

#>           filename cpred   spc.405 spc.405.5   spc.406 spc.406.5   spc.407
#> 1 rawdata/flu2.txt     1  66.80133  63.71533  66.71200  69.58233  72.52967
#> 2 rawdata/flu5.txt     3 167.26667 171.89833 177.47067 184.62467 189.75233
#> 3 rawdata/flu6.txt     2 198.43033 209.45800 215.78500 224.58700 232.52800
#> 4 rawdata/flu6.txt     4 198.43033 209.45800 215.78500 224.58700 232.52800
#>   spc.407.5   spc.408 spc.408.5   spc.409 spc.409.5   spc.410 spc.410.5
#> 1  74.55833  77.04800  80.25967  82.53867  84.49167  88.15167  91.08533
#> 2 198.11533 205.56267 208.41933 217.55267 224.74633 231.03567 236.35233
#> 3 240.77133 248.04667 256.89133 262.73900 270.27133 281.82867 287.70467
#> 4 240.77133 248.04667 256.89133 262.73900 270.27133 281.82867 287.70467
#>     spc.411 spc.411.5   spc.412 spc.412.5   spc.413 spc.413.5   spc.414
#> 1  95.37233  95.53033  98.99533 101.03433 103.55800 107.02667 109.54533
#> 2 244.61233 251.95000 257.34633 265.54500 272.06333 280.74133 288.10567
#> 3 297.23500 307.59333 314.99367 324.56300 330.80033 340.89667 347.62167
#> 4 297.23500 307.59333 314.99367 324.56300 330.80033 340.89667 347.62167
#>   spc.414.5   spc.415 spc.415.5   spc.416 spc.416.5   spc.417 spc.417.5
#> 1 111.67233 113.98733 117.03267 119.83500 123.49433 123.32367 127.00167
#> 2 291.88367 300.74000 307.87567 315.42967 321.89600 326.73133 334.01200
#> 3 355.02667 365.62333 374.35700 380.21733 391.40100 398.17300 405.42933
#> 4 355.02667 365.62333 374.35700 380.21733 391.40100 398.17300 405.42933
#>     spc.418 spc.418.5   spc.419 spc.419.5   spc.420 spc.420.5   spc.421
#> 1 130.68067 132.35800 136.16467 138.39933 140.22433 141.04333 143.13133
#> 2 339.59933 343.86167 351.25667 357.79733 360.36000 367.96533 374.54900
#> 3 412.48400 421.27300 429.36267 433.35767 443.18733 447.88033 461.33933
#> 4 412.48400 421.27300 429.36267 433.35767 443.18733 447.88033 461.33933
#>   spc.421.5   spc.422 spc.422.5   spc.423 spc.423.5   spc.424 spc.424.5
#> 1 147.17967 150.01200 153.69333 155.72800 158.47200 159.50067 160.81567
#> 2 381.76000 389.24100 397.79733 403.78367 408.02300 418.35533 420.61400
#> 3 467.12733 475.18967 481.66667 488.73567 494.01867 504.47333 515.40467
#> 4 467.12733 475.18967 481.66667 488.73567 494.01867 504.47333 515.40467
#>     spc.425 spc.425.5   spc.426 spc.426.5   spc.427 spc.427.5   spc.428
#> 1 163.90967 166.25500 169.59233 170.82833 172.08767 175.21133 178.47067
#> 2 427.04333 433.33900 438.43633 449.57667 449.64833 455.43800 463.33200
#> 3 521.23867 522.64633 536.03533 544.41767 551.63333 556.71267 567.61267
#> 4 521.23867 522.64633 536.03533 544.41767 551.63333 556.71267 567.61267
#>   spc.428.5   spc.429 spc.429.5   spc.430 spc.430.5   spc.431 spc.431.5
#> 1 179.68100 181.82867 184.42933 187.75600 187.93767 191.05700 192.50900
#> 2 471.47433 476.60033 480.93533 486.53833 491.91000 500.59200 502.12433
#> 3 572.54933 578.72667 587.74900 590.54867 600.42300 608.35800 614.03833
#> 4 572.54933 578.72667 587.74900 590.54867 600.42300 608.35800 614.03833
#>     spc.432 spc.432.5   spc.433 spc.433.5   spc.434 spc.434.5   spc.435
#> 1 192.80100 195.55600 197.89067 198.78200 200.99467 201.52233 200.20600
#> 2 508.40733 512.70433 515.62733 522.20267 524.72467 525.20700 529.69967
#> 3 616.95567 621.87267 629.80033 633.44500 637.13233 639.16167 642.89833
#> 4 616.95567 621.87267 629.80033 633.44500 637.13233 639.16167 642.89833
#>   spc.435.5   spc.436 spc.436.5   spc.437 spc.437.5   spc.438 spc.438.5
#> 1 204.13200 204.90433 203.98833 205.75600 204.81700 206.09100 208.60733
#> 2 531.71400 537.00600 538.20567 543.83567 544.08533 547.93633 552.15700
#> 3 648.17200 651.67900 654.53967 656.83267 659.77800 662.05633 663.12300
#> 4 648.17200 651.67900 654.53967 656.83267 659.77800 662.05633 663.12300
#>     spc.439 spc.439.5   spc.440 spc.440.5   spc.441 spc.441.5   spc.442
#> 1 210.79667 210.58400 210.19033 210.54233 209.85933 212.82733 212.52067
#> 2 550.93500 548.20333 548.14767 550.77700 550.86067 551.78767 553.30400
#> 3 663.48200 670.47567 669.88700 670.06033 671.00800 667.61233 666.30833
#> 4 663.48200 670.47567 669.88700 670.06033 671.00800 667.61233 666.30833
#>   spc.442.5   spc.443 spc.443.5   spc.444 spc.444.5   spc.445 spc.445.5
#> 1 212.41967 211.22233 215.56367 213.42467 211.15933 212.97533 213.98933
#> 2 552.57400 553.54533 551.11500 555.87400 554.84800 558.49500 556.92500
#> 3 669.97433 675.68300 675.97867 676.64200 674.73667 671.23433 673.61133
#> 4 669.97433 675.68300 675.97867 676.64200 674.73667 671.23433 673.61133
#>     spc.446 spc.446.5   spc.447 spc.447.5   spc.448 spc.448.5   spc.449
#> 1 214.17200 213.47633 212.48867 211.74100 212.70467 212.94033 212.12933
#> 2 559.01233 557.12833 555.68833 553.09200 554.96233 553.30000 555.83300
#> 3 676.73533 677.49467 674.28033 676.64567 674.89067 673.09200 674.89800
#> 4 676.73533 677.49467 674.28033 676.64567 674.89067 673.09200 674.89800
#>   spc.449.5   spc.450 spc.450.5   spc.451 spc.451.5   spc.452 spc.452.5
#> 1 213.77267 213.49733 213.46400 213.17067 212.83600 211.96333 208.79867
#> 2 555.17800 556.52367 554.77567 551.07633 553.91567 553.45200 550.76067
#> 3 671.64167 672.53167 672.28033 671.00033 672.87900 674.74300 672.72467
#> 4 671.64167 672.53167 672.28033 671.00033 672.87900 674.74300 672.72467
#>     spc.453 spc.453.5   spc.454 spc.454.5   spc.455 spc.455.5   spc.456
#> 1 211.50633 209.47700 211.85967 213.26200 212.28367 211.77333 209.39067
#> 2 551.41767 548.26100 549.39067 553.83833 549.54067 544.38700 544.44033
#> 3 669.79100 672.00433 672.33967 668.74033 662.63333 666.29700 664.63267
#> 4 669.79100 672.00433 672.33967 668.74033 662.63333 666.29700 664.63267
#>   spc.456.5   spc.457 spc.457.5   spc.458 spc.458.5   spc.459 spc.459.5
#> 1 208.85600 208.34033 206.50633 206.77733 206.64500 205.25500 201.74700
#> 2 543.01167 539.73833 538.81333 537.68600 532.22333 531.40200 527.58867
#> 3 659.91800 660.13033 656.23733 654.19500 652.31833 647.82100 643.38133
#> 4 659.91800 660.13033 656.23733 654.19500 652.31833 647.82100 643.38133
#>     spc.460 spc.460.5   spc.461 spc.461.5   spc.462 spc.462.5   spc.463
#> 1 201.98800 199.57000 200.55100 198.76767 197.03300 194.14467 193.38300
#> 2 526.91933 519.80200 518.62067 515.22900 511.55500 506.87467 503.26400
#> 3 640.79033 632.11967 633.56367 627.81900 623.21133 619.99667 613.55133
#> 4 640.79033 632.11967 633.56367 627.81900 623.21133 619.99667 613.55133
#>   spc.463.5   spc.464 spc.464.5   spc.465 spc.465.5   spc.466 spc.466.5
#> 1 193.66467 192.51267 190.63300 186.45567 186.05433 182.30533 182.33367
#> 2 498.34400 494.06167 491.78467 484.60533 486.67233 479.56767 475.90100
#> 3 606.01267 605.92100 599.88400 593.00650 587.40150 586.38900 581.49850
#> 4 606.01267 605.92100 599.88400 593.00650 587.40150 586.38900 581.49850
#>     spc.467 spc.467.5   spc.468 spc.468.5   spc.469 spc.469.5   spc.470
#> 1 180.56933 180.50633 180.21267 175.21100 171.65000 170.41667 170.37400
#> 2 472.13100 464.54333 457.90900 455.22167 449.28400 446.34567 442.75167
#> 3 574.15100 566.49250 560.29000 559.46250 550.27850 545.54350 542.76900
#> 4 574.15100 566.49250 560.29000 559.46250 550.27850 545.54350 542.76900
#>   spc.470.5   spc.471 spc.471.5   spc.472 spc.472.5   spc.473 spc.473.5
#> 1 169.40700 164.89000 164.14133 163.17333 160.98933 159.66533 157.74633
#> 2 438.25067 432.26667 427.76167 422.73000 416.02133 413.34567 408.62267
#> 3 528.92800 528.12900 519.65350 516.54950 512.45650 502.48600 499.65250
#> 4 528.92800 528.12900 519.65350 516.54950 512.45650 502.48600 499.65250
#>     spc.474 spc.474.5   spc.475 spc.475.5   spc.476 spc.476.5   spc.477
#> 1 155.13467 153.23933 151.52300 149.19667 147.49467 147.49000 145.43367
#> 2 403.80533 400.32500 393.77333 393.78567 386.53867 383.86933 378.65500
#> 3 495.00100 486.49700 481.73950 476.47200 472.37900 469.25450 466.99550
#> 4 495.00100 486.49700 481.73950 476.47200 472.37900 469.25450 466.99550
#>   spc.477.5   spc.478 spc.478.5   spc.479 spc.479.5   spc.480 spc.480.5
#> 1 144.81900 142.43867 142.68967 142.88100 139.83767 135.47933 135.25267
#> 2 373.64800 372.70233 368.87433 366.15533 362.89967 357.86333 354.96700
#> 3 457.76300 452.32300 450.48350 445.14500 444.27400 435.22800 430.33300
#> 4 457.76300 452.32300 450.48350 445.14500 444.27400 435.22800 430.33300
#>     spc.481 spc.481.5   spc.482 spc.482.5   spc.483 spc.483.5   spc.484
#> 1 135.84567 133.55900 133.31767 132.82400 128.93033 125.20300 124.73967
#> 2 352.37067 344.31567 342.10133 339.01433 336.71533 334.59300 326.09833
#> 3 424.72000 420.74800 418.46650 414.01750 408.32400 403.85400 405.00450
#> 4 424.72000 420.74800 418.46650 414.01750 408.32400 403.85400 405.00450
#>   spc.484.5   spc.485 spc.485.5   spc.486 spc.486.5   spc.487 spc.487.5
#> 1 124.71667 121.67833 120.90167 121.40600 117.66767 117.04100 115.94100
#> 2 324.69400 320.56700 315.93067 312.73200 307.94367 306.38633 298.34667
#> 3 394.33500 389.97750 384.49100 382.33700 375.08200 369.09900 366.06300
#> 4 394.33500 389.97750 384.49100 382.33700 375.08200 369.09900 366.06300
#>     spc.488 spc.488.5   spc.489 spc.489.5   spc.490 spc.490.5   spc.491
#> 1 112.22900 112.72433 111.67267 109.69833 107.64067 108.05067 105.22000
#> 2 298.57433 296.05233 292.49933 285.46933 282.62833 279.92767 273.34667
#> 3 362.45350 358.99650 352.14300 348.20000 347.10650 338.79150 335.34700
#> 4 362.45350 358.99650 352.14300 348.20000 347.10650 338.79150 335.34700
#>   spc.491.5   spc.492 spc.492.5   spc.493 spc.493.5   spc.494 spc.494.5
#> 1 102.95300 102.75833 100.62800  97.96500  97.35333  96.60167  96.20600
#> 2 267.55433 266.10100 261.76567 259.40200 256.20067 252.06567 248.06700
#> 3 330.04000 325.16200 322.06850 315.64550 312.32950 307.51850 302.32550
#> 4 330.04000 325.16200 322.06850 315.64550 312.32950 307.51850 302.32550
#>     spc.495    c
#> 1  94.61033 0.10
#> 2 246.95200 0.25
#> 3 294.64950 0.30
#> 4 294.64950 0.30