Convert a hyperSpec object to character strings for Display print, show, and summary show the result of as.character.

print, show, and summary differ only in the defaults. show displays the range of values instead,

# S4 method for hyperSpec
as.character(
  x,
  digits = getOption("digits"),
  range = TRUE,
  max.print = 5,
  shorten.to = c(2, 1)
)

# S4 method for hyperSpec
show(object)

# S4 method for hyperSpec
print(x, range = FALSE, ...)

# S4 method for hyperSpec
summary(object, ...)

Arguments

x
a hyperSpec object

digits
number of digits handed over to format

range
should the values be indicated as range rather then first and last elements?

max.print
maximum number of elements to be printed (of a variable)

shorten.to
if a vector is longer than max.print, only the first shorten.to[1] and the last shorten.to[2] elements are printed

object
a hyperSpec object

...
print and summary hand further arguments to as.character

Value

as.character returns a character vector fit to be printed by cat with sep = "\n".

print invisibly returns x after printing, show returns an invisible NULL.

See also

Examples

chondro

#> hyperSpec object
#>    875 spectra
#>    5 data columns
#>    300 data points / spectrum
#> wavelength: Delta * tilde(nu)/cm^-1 [numeric] 602 606 ... 1798 
#> data:  (875 rows x 5 columns)
#>    1. y: y [numeric] -4.77 -4.77 ... 19.23 
#>    2. x: x [numeric] -11.55 -10.55 ... 22.45 
#>    3. filename: filename [character] rawdata/chondro.txt rawdata/chondro.txt ... rawdata/chondro.txt 
#>    4. clusters: clusters [factor] matrix matrix ... lacuna + NA
#>    5. spc: I / a.u. [matrix, array300] 501.8194 500.4552 ... 169.2942 

show (chondro)

#> hyperSpec object
#>    875 spectra
#>    5 data columns
#>    300 data points / spectrum
#> wavelength: Delta * tilde(nu)/cm^-1 [numeric] 602 606 ... 1798 
#> data:  (875 rows x 5 columns)
#>    1. y: y [numeric] -4.77 -4.77 ... 19.23 
#>    2. x: x [numeric] -11.55 -10.55 ... 22.45 
#>    3. filename: filename [character] rawdata/chondro.txt rawdata/chondro.txt ... rawdata/chondro.txt 
#>    4. clusters: clusters [factor] matrix matrix ... lacuna + NA
#>    5. spc: I / a.u. [matrix, array300] 501.8194 500.4552 ... 169.2942 

summary (chondro)

#> hyperSpec object
#>    875 spectra
#>    5 data columns
#>    300 data points / spectrum
#> wavelength: Delta * tilde(nu)/cm^-1 [numeric] 602 606 ... 1798 
#> data:  (875 rows x 5 columns)
#>    1. y: y [numeric] -4.77 -4.77 ... 19.23 
#>    2. x: x [numeric] -11.55 -10.55 ... 22.45 
#>    3. filename: filename [character] rawdata/chondro.txt rawdata/chondro.txt ... rawdata/chondro.txt 
#>    4. clusters: clusters [factor] matrix matrix ... lacuna + NA
#>    5. spc: I / a.u. [matrix, array300] 501.8194 500.4552 ... 169.2942 

print (chondro, range = TRUE)

#> hyperSpec object
#>    875 spectra
#>    5 data columns
#>    300 data points / spectrum
#> wavelength: Delta * tilde(nu)/cm^-1 [numeric] 602 606 ... 1798 
#> data:  (875 rows x 5 columns)
#>    1. y: y [numeric] -4.77 -4.77 ... 19.23 
#>    2. x: x [numeric] -11.55 -10.55 ... 22.45 
#>    3. filename: filename [character] rawdata/chondro.txt rawdata/chondro.txt ... rawdata/chondro.txt 
#>    4. clusters: clusters [factor] matrix matrix ... lacuna + NA
#>    5. spc: I / a.u. [matrix, array300] 501.8194 500.4552 ... 169.2942